% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AStest.R
\name{AStest}
\alias{AStest}
\title{This function computes the AS test using DGM implementation in the package RationalExp}
\usage{
AStest(lamb, YY, XX, tuningParam = NULL)
}
\arguments{
\item{lamb}{the point under the form lambda q to be tested.}

\item{YY}{the observations of the outcome variable.}

\item{XX}{the observations of the regressor X'q variable.}

\item{tuningParam}{the list of tuning parameters. For the details see the function "test" in the package RationalExp.}
}
\value{
the result of the test at level 5%
}
\description{
This function computes the AS test using DGM implementation in the package RationalExp
}
