% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_constraints.R
\name{compute_constraints}
\alias{compute_constraints}
\title{Compute the indexes of the values of the common regressors Xc used in the various shape constraints}
\usage{
compute_constraints(
  constraint,
  values,
  values_sel,
  indexes_k = NULL,
  nbV,
  grouped0,
  ind = NULL,
  c_sign = NULL
)
}
\arguments{
\item{constraint}{the current shape constraint}

\item{values}{the different unique points of support of the common regressor Xc.}

\item{values_sel}{the selected values of Xc for the conditioning. Default is NULL.}

\item{indexes_k}{indexes of the constraints}

\item{nbV}{indexes of the constraints}

\item{grouped0}{boolean indexing if the values of Xc have been changed}

\item{ind}{index}

\item{c_sign}{sign restrictions on the commonly observed regressors: -1 for a minus sign, 1 for a plus sign, 0 otherwise. Default is NULL, i.e. no constraints.}
}
\value{
a vector containing:

- the matrix R where each line is a constraint

- the matrices pp0 and pp1, which contains the indexes of the values of Xc in values_sel which enters the various constraints.
}
\description{
Compute the indexes of the values of the common regressors Xc used in the various shape constraints
}
