% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_radial.R
\name{compute_radial}
\alias{compute_radial}
\title{Function to compute the DGM bounds on the noncommon regressor Xnc}
\usage{
compute_radial(
  sample1 = NULL,
  Xc_x,
  Xnc,
  Xc_y,
  Y,
  values,
  dimXc,
  dimXnc,
  nb_pts,
  sam0,
  kp0,
  data_k = NULL,
  Opt = "boot",
  lim = 1,
  weights_x = NULL,
  weights_y = NULL,
  c_sign = NULL,
  nc_sign = NULL,
  refs0 = NULL,
  type = "both",
  meth = "adapt",
  trunc = 2,
  boot_par = 0.35,
  winsor = winsor,
  version = "first",
  ratio_ref = NULL,
  ratio_ref_m = NULL,
  Dmat_Yk_ref = NULL,
  Dmat_Xk_ref = NULL,
  grid_I_ref = NULL,
  es_all = NULL,
  T_xy_all = NULL
)
}
\arguments{
\item{sample1}{if NULL compute the point estimate, if a natural number then evaluate a bootstrap or subsampling replication.}

\item{Xc_x}{the common regressor on the dataset  (Xnc,Xc). Default is NULL.}

\item{Xnc}{the noncommon regressor on the dataset  (Xnc,Xc). No default.}

\item{Xc_y}{the common regressor on the dataset  (Y,Xc). Default is NULL.}

\item{Y}{the outcome variable. No default.}

\item{values}{the different unique points of support of the common regressor Xc.}

\item{dimXc}{the dimension of the common regressors Xc.}

\item{dimXnc}{the dimension of the noncommon regressors Xnc.}

\item{nb_pts}{the constant C in DGM for the epsilon_0, the lower bound on the grid for epsilon, taken equal to nb_pts*ln(n)/n. Default is 1 without regressors Xc, 3 with Xc.}

\item{sam0}{the directions q to compute the variance bounds on the radial function.}

\item{kp0}{the matrix containing the directions q and the selected epsilon(q).}

\item{data_k}{the number of points for the grid search on epsilon. Default is 30. If NULL, then epsilon is taken fixed equal to kp.}

\item{Opt}{the option of the method to compute the critical values, either "boot" (numerical bootstrap) or "subsampling". Default is numerical bootstrap.}

\item{lim}{the limit number of observations under which we do no compute the conditional variance.}

\item{weights_x}{the sampling weights for the dataset (Xnc,Xc).}

\item{weights_y}{the sampling weights for the dataset (Y,Xc).}

\item{c_sign}{sign restrictions on the commonly observed regressors: -1 for a minus sign, 1 for a plus sign, 0 otherwise. Default is NULL, i.e. no contraints.}

\item{nc_sign}{sign restrictions on the non-commonly observed regressors Xnc: -1 for a minus sign, 1 for a plus sign, 0 otherwise. Default is NULL, i.e. no contraints.}

\item{refs0}{indicating the positions in the vector values corresponding to the components of betac.}

\item{type}{Equal to "both".}

\item{meth}{the method for the choice of epsilon, either "adapt", i.e. adapted to the direction or "min" the minimum over the directions. Default is "adapt".}

\item{trunc}{equal to 2, for the definition of epsilon.}

\item{boot_par}{the numerical bootstrap parameter. Default is 0.3 without Xc, 0.35 with Xc.}

\item{winsor}{indicates if winsorisation. Default is FALSE.}

\item{version}{version of the computation of the ratio, "first" is a degraded version but fast; "second" is a correct version but slower. Default is "second".}

\item{ratio_ref}{the point estimate of the ratio R in the radial function.}

\item{ratio_ref_m}{the point estimate of the ratio R in the radial function on the opposite direction.}

\item{Dmat_Yk_ref}{the vector of means of Y| Xc on the initial dataset.}

\item{Dmat_Xk_ref}{the vector of means of Xnc| Xc on the initial dataset.}

\item{grid_I_ref}{the grid of alpha on which we evaluate the ratio R to compute the point estimate of the radial function.}

\item{es_all}{the numerical bootstrap parameter delta_n}

\item{T_xy_all}{the apparent sample size the taking into account the difference in the two datasets.}
}
\value{
a list contaning:

 - upper: the upper bound in the specified directions, possibly with sign constraints

 - lower: the lower bound in the specified directions, possibly with sign constraints

 - unconstr: the bounds without sign constraints in the specified directions

 - Ykmean: the means of Y|Xc for the considered sample

 - Xkmean: the means of Xnc|Xc for the considered sample

 - DYk: the difference of means of Y|Xc =k -  Y|Xc =0 for the considered sample

 - DXk: the difference of means of Xnc|Xc =k -  Xnc|Xc =0 for the considered sample

 - tests: the pvalues of the tests H0 : DXk =0

 - ratio_ref: the ratio R in the radial function computed for the initial sample
}
\description{
Function to compute the DGM bounds on the noncommon regressor Xnc
}
