% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_ratio_point.R
\name{compute_ratio_point}
\alias{compute_ratio_point}
\title{Function to compute the main statistic for the point estimate}
\usage{
compute_ratio_point(
  x_kp0,
  Xp,
  Yp,
  for_critY,
  dimXnc,
  weights_xp,
  weights_yp,
  T_xy1,
  T_xy,
  es,
  version = "first",
  trunc = 2,
  grid_I = NULL,
  winsor = FALSE
)
}
\arguments{
\item{x_kp0}{a matrix containing the directions to compute the radial function, and the associated choice epsilon(q).}

\item{Xp}{the observations of the noncommon regressor (possibly conditional on Xc).}

\item{Yp}{the observations of the outcome variable.}

\item{for_critY}{the numerator of the ratio R for the point estimate of the radial function, on the grid grid_I;}

\item{dimXnc}{the dimension of the noncommon regressors}

\item{weights_xp}{the sampling or bootstrap weights for the dataset (Xnc,Xc).}

\item{weights_yp}{the sampling or bootstrap weights for the dataset (Y,Xc).}

\item{T_xy1}{the apparent sample size taking into account the difference in the two datasets.}

\item{T_xy}{the apparent sample size, possibly conditional on Xc, the taking into account the difference in the two datasets.}

\item{es}{the numerical bootstrap parameter delta_n.}

\item{version}{version of the computation of the ratio, "first" is a degraded version but fast; "second" is a correct version but slower. Default is "second".}

\item{trunc}{equal to 2, for the definition of epsilon.}

\item{grid_I}{the grid of alpha on which we evaluate the ratio R to compute the point estimate of the radial function.}

\item{winsor}{indicates if winsorisation. Default is FALSE.}
}
\value{
the value of the point estimate of the radial function using the DGM method.
}
\description{
Function to compute the main statistic for the point estimate
}
