% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_stat_variance.R
\name{compute_stat_variance}
\alias{compute_stat_variance}
\title{Function to compute the Variance bounds on the noncommon regressor Xnc}
\usage{
compute_stat_variance(
  sample1 = NULL,
  X1_x,
  X2,
  X1_y,
  Y,
  values,
  refs0,
  dimX1,
  dimX2,
  nb_pts,
  sam0,
  kp,
  Opt = "boot",
  lim = 1,
  weights_x = NULL,
  weights_y = NULL,
  boot_par = 0.35,
  c_sign = NULL,
  nc_sign = NULL,
  ratio_ref = NULL,
  Dmat_Yk_ref = NULL,
  Dmat_Xk_ref = NULL
)
}
\arguments{
\item{sample1}{if NULL compute the point estimate, if a natural number then evaluate a bootstrap or subsampling replication.}

\item{X1_x}{the common regressor on the dataset  (Xnc,Xc). Default is NULL.}

\item{X2}{the noncommon regressor on the dataset  (Xnc,Xc). No default.}

\item{X1_y}{the common regressor on the dataset  (Y,Xc). Default is NULL.}

\item{Y}{the outcome variable. No default.}

\item{values}{the different unique points of support of the common regressor Xc.}

\item{refs0}{indicating the positions in the vector values corresponding to the components of betac.}

\item{dimX1}{the dimension of the common regressors Xc.}

\item{dimX2}{the dimension of the noncommon regressors Xnc.}

\item{nb_pts}{the constant C in DGM for the epsilon_0, the lower bound on the grid for epsilon, taken equal to nb_pts*ln(n)/n. Default is 1 without regressors Xc, 3 with Xc.}

\item{sam0}{the directions q to compute the variance bounds on the radial function.}

\item{kp}{If data_k =NULL, then epsilon is taken equal to kp.}

\item{Opt}{the option of the method to compute the critical values, either "boot" (numerical bootstrap) or "subsampling". Default is numerical bootstrap.}

\item{lim}{the limit number of observations under which we do no compute the conditional variance.}

\item{weights_x}{the sampling weights for the dataset (Xnc,Xc).}

\item{weights_y}{the sampling weights for the dataset (Y,Xc).}

\item{boot_par}{the numerical bootstrap parameter. Default is 0.3 without Xc, 0.35 with Xc.}

\item{c_sign}{sign restrictions on the commonly observed regressors: -1 for a minus sign, 1 for a plus sign, 0 otherwise. Default is NULL, i.e. no constraints.}

\item{nc_sign}{sign restrictions on the non-commonly observed regressors Xnc: -1 for a minus sign, 1 for a plus sign, 0 otherwise. Default is NULL, i.e. no constraints.}

\item{ratio_ref}{the variances ratio of the point estimate for the computation of the numerical boostrap replications. Default is NULL.}

\item{Dmat_Yk_ref}{the vector of means of Y| Xc on the initial dataset.}

\item{Dmat_Xk_ref}{the vector of means of Xnc| Xc on the initial dataset.}
}
\value{
a list containing:

 - upper: the upper bound in the specified directions, possibly with sign constraints

 - lower: the lower bound in the specified directions, possibly with sign constraints

 - unconstr: the bounds without sign constraints in the specified directions

 - Ykmean: the means of Y|Xc for the considered sample

 - Xkmean: the means of Xnc|Xc for the considered sample

 - DYk: the difference of means of Y|Xc =k -  Y|Xc =0 for the considered sample

 - DXk: the difference of means of Xnc|Xc =k -  Xnc|Xc =0 for the considered sample

 - tests: the pvalues of the tests H0 : DXk =0

 - ratio_ref: the ratio R in the radial function computed for the initial sample
}
\description{
Function to compute the Variance bounds on the noncommon regressor Xnc
}
