% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_support_paral.R
\name{compute_support_paral}
\alias{compute_support_paral}
\title{Function to minimize to compute the function sigma for the projections of the identified set}
\usage{
compute_support_paral(
  dir_nb,
  sam0,
  Xnc,
  kp0,
  data_k,
  dimXc,
  dimXnc,
  Xc_xb = NULL,
  Xncb,
  Xc_yb = NULL,
  Yb,
  values,
  Opt,
  weights_x,
  weights_y,
  c_sign,
  nc_sign,
  refs0,
  meth,
  boot_par,
  T_xy,
  bc,
  winsor,
  version,
  weights_xs,
  weights_ys
)
}
\arguments{
\item{dir_nb}{the reference for the considered direction e in sam0}

\item{sam0}{the directions q to compute the radial function.}

\item{Xnc}{the noncommon regressor on the dataset  (Xnc,Xc). No default}

\item{kp0}{the matrix containing the directions q and the selected epsilon(q)}

\item{data_k}{the number of points for the grid search on epsilon. Default is 30. If NULL, then epsilon is taken fixed equal to kp.}

\item{dimXc}{the dimension of the common regressors Xc.}

\item{dimXnc}{the dimension of the noncommon regressors Xnc.}

\item{Xc_xb}{the possibly bootstraped/subsampled common regressor on the dataset  (Xnc,Xc). Default is NULL.}

\item{Xncb}{the possibly bootstraped/subsampled noncommon regressor on the dataset (Xnc,Xc). No default.}

\item{Xc_yb}{the possibly bootstraped/subsampled common regressor on the dataset  (Y,Xc). Default is NULL.}

\item{Yb}{the possibly bootstraped/subsampled outcome variable  on the dataset  (Y,Xc). No default.}

\item{values}{the different unique points of support of the common regressor Xc.}

\item{Opt}{the option of the method to compute the critical values, either "boot" (numerical bootstrap) or "subsampling". Default is numerical bootstrap.}

\item{weights_x}{the bootstrap or sampling weights for the dataset (Xnc,Xc).}

\item{weights_y}{the bootstrap or sampling weights for the dataset (Y,Xc).}

\item{c_sign}{sign restrictions on the commonly observed regressors: -1 for a minus sign, 1 for a plus sign, 0 otherwise. Default is NULL, i.e. no constraints.}

\item{nc_sign}{sign restrictions on the non-commonly observed regressors Xnc: -1 for a minus sign, 1 for a plus sign, 0 otherwise. Default is NULL, i.e. no constraints.}

\item{refs0}{indicating the positions in the vector values corresponding to the components of betac.}

\item{meth}{the method for the choice of epsilon, either "adapt", i.e. adapted to the direction or "min" the minimum over the directions. Default is "adapt".}

\item{boot_par}{the numerical bootstrap parameter. Default is 0.3 without Xc, 0.35 with Xc.}

\item{T_xy}{the apparent sample size the taking into account the difference in the two datasets.}

\item{bc}{if TRUE compute also the bounds on betac. Default is FALSE.}

\item{winsor}{indicates if winsorisation. Default is FALSE.}

\item{version}{version of the computation of the ratio, "first" is a degraded version but fast; "second" is a correct version but slower. Default is "second".}

\item{weights_xs}{the sampling weights for the dataset (Xnc,Xc).}

\item{weights_ys}{the bootstrap or sampling weights for the dataset (Y,Xc).}
}
\value{
the value of the support function in the specifed direction dir_nb.
}
\description{
Function to minimize to compute the function sigma for the projections of the identified set
}
