% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_values.R
\name{create_values}
\alias{create_values}
\title{Function to create the matrix of the support points for the common regressors Xc}
\usage{
create_values(dimX, c_var, Rdata)
}
\arguments{
\item{dimX}{the dimension of the common regressors Xc.}

\item{c_var}{the label of these regressors.}

\item{Rdata}{dataset containing (Xnc,Xc) where Xnc are the non commonly observed regressors, Xc are potential common regressors.}
}
\value{
a matrix of the values of the support points for the common regressors Xc
}
\description{
Function to create the matrix of the support points for the common regressors Xc
}
