% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DGM_bounds.R
\name{DGM_bounds}
\alias{DGM_bounds}
\title{This function compute the DGM bounds for all the different coefficients.}
\usage{
DGM_bounds(
  Ldata,
  Rdata,
  values,
  sam0,
  refs0,
  out_var,
  nc_var,
  c_var = NULL,
  nc_sign = NULL,
  c_sign = NULL,
  nbCores = 1,
  kp = 0.5,
  nb_pts = 1,
  Opt = "boot",
  Bsamp = 1000,
  data_k = 30,
  C0 = 0.5,
  list_ex = c(),
  weights_x = NULL,
  weights_y = NULL,
  outside = FALSE,
  trunc = 2,
  boot_par = 0.3,
  meth = "adapt",
  modeNA = FALSE,
  winsor = FALSE,
  version = "second",
  alpha = 0.05,
  projections = FALSE
)
}
\arguments{
\item{Ldata}{dataset containing (Y,Xc) where Y is the outcome, Xc are potential common regressors.}

\item{Rdata}{dataset containing (Xnc,Xc) where Xnc are the non commonly observed regressors, Xc are potential common regressors.}

\item{values}{the different unique points of support of the common regressor Xc.}

\item{sam0}{the directions q to compute the radial function.}

\item{refs0}{indicating the positions in the vector values corresponding to the components of betac.}

\item{out_var}{label of the outcome variable Y.}

\item{nc_var}{label of the non commonly observed regressors Xnc.}

\item{c_var}{label of the commonly observed regressors Xc.}

\item{nc_sign}{sign restrictions on the non-commonly observed regressors Xnc: -1 for a minus sign, 1 for a plus sign, 0 otherwise. Default is NULL, i.e. no constraints.}

\item{c_sign}{sign restrictions on the commonly observed regressors: -1 for a minus sign, 1 for a plus sign, 0 otherwise. Default is NULL, i.e. no constraints.}

\item{nbCores}{number of cores for the parallel computation. Default is 1.}

\item{kp}{If data_k =NULL, then epsilon is taken equal to kp.}

\item{nb_pts}{the constant C in DGM for the epsilon_0, the lower bound on the grid for epsilon, taken equal to nb_pts*ln(n)/n. Default is 1 without regressors Xc, 3 with Xc.}

\item{Opt}{the option of the method to compute the critical values, either "boot" (numerical bootstrap) or "subsampling". Default is numerical bootstrap.}

\item{Bsamp}{the number of bootstrap/subsampling replications. Default is 1000.}

\item{data_k}{the number of points for the grid search on epsilon. Default is 30. If NULL, then epsilon is taken fixed equal to kp.}

\item{C0}{the upper bound on the grid for epsilon. Default is 0.5.}

\item{list_ex}{the list of the data that should not be used on the different clusters. This reduces the computational cost. Default is empty list.}

\item{weights_x}{the sampling weights for the dataset (Xnc,Xc). Default is NULL.}

\item{weights_y}{the sampling weights for the dataset (Y,Xc). Default is NULL.}

\item{outside}{if TRUE indicates that the parallel computing has been launched outside of the function. Default is FALSE.}

\item{trunc}{equal to 2, for the definition of epsilon.}

\item{boot_par}{the numerical bootstrap parameter. Default is 0.3 without Xc, 0.35 with Xc.}

\item{meth}{the method for the choice of epsilon, either "adapt", i.e. adapted to the direction or "min" the minimum over the directions. Default is "adapt".}

\item{modeNA}{indicates if NA introduced if the interval is empty. Default is FALSE.}

\item{winsor}{indicates if winsorisation. Default is FALSE.}

\item{version}{version of the computation of the ratio, "first" is a degraded version but fast; "second" is a correct version but slower. Default is "second".}

\item{alpha}{for the level of the confidence region. Default is 0.05.}

\item{projections}{if FALSE compute the identified set along some directions or the confidence regions. Default is FALSE}
}
\value{
a list containing, in order:
 - ci : a list with all the information on the confidence intervals

    * upper: upper bound of the confidence interval on the radial function S in the specified direction at level alpha, possibly with sign constraints

    * lower: lower bound upper bound of the confidence interval on the radial function S, possibly with sign constraints

    * unconstr: confidence interval on the radial function S, without sign constraints

    * betac_ci: confidence intervals on each coefficients related to the common regressor, possibly with sign constraints

    * betac_ci_unc: confidence intervals on each coefficients related to the common regressor without sign constraints

    If projection is TRUE:

       * support: confidence bound on the support function in each specified direction

 - point : a list with all the information on the point estimates

    * upper: the upper bounds on betanc, possibly with sign constraints

    * lower: the lower bounds on betanc, possibly with sign constraints

    * unconstr: bounds on betanc without sign constraints

    * betac_pt: bounds on betanc, possibly with sign constraints

    * betac_pt_unc: bounds on betanc without sign constraints
     If projection ==TRUE:

    * support: point estimate of the support function in each specified direction

 - epsilon : the values of the selected epsilon(q)
}
\description{
This function compute the DGM bounds for all the different coefficients.
}
