% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ewcdf.R
\name{ewcdf}
\alias{ewcdf}
\title{Compute the weighted empirical cumulative distribution}
\usage{
ewcdf(x, weights = rep(1/length(x), length(x)))
}
\arguments{
\item{x}{the sample}

\item{weights}{the associated weights if any. Default is uniform.}
}
\value{
a vector containing:

 - the weighted empirical cumulative distribution function

 - the cumulated weights associated to the ordered values of the random variable.
}
\description{
Compute the weighted empirical cumulative distribution
}
