% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_epsilon.R
\name{select_epsilon}
\alias{select_epsilon}
\title{Function for the data-driven selection of the epsilon tuning parameter}
\usage{
select_epsilon(
  sam1,
  kp,
  Xc_x,
  Xnc,
  Xc_y,
  Y,
  values,
  dimXc,
  dimXnc,
  nb_pts,
  Opt,
  lim,
  weights_x,
  weights_y,
  refs0,
  trunc,
  boot_par = 0.35,
  data_k = 30,
  C0 = 0.5,
  c_sign = NULL,
  nc_sign = NULL,
  meth = "adapt",
  nbCores = 1,
  winsor = FALSE,
  version = "first",
  alpha = 0.05
)
}
\arguments{
\item{sam1}{the matrix containing the directions q on which to compute the selected rule for epsilon(q)}

\item{kp}{If data_k =NULL, then epsilon is taken equal to kp.}

\item{Xc_x}{the common regressor on the dataset  (Xnc,Xc). Default is NULL.}

\item{Xnc}{the noncommon regressor on the dataset  (Xnc,Xc). No default.}

\item{Xc_y}{the common regressor on the dataset  (Y,Xc). Default is NULL.}

\item{Y}{the outcome variable. No default.}

\item{values}{the different unique points of support of the common regressor Xc.}

\item{dimXc}{the dimension of the common regressors Xc.}

\item{dimXnc}{the dimension of the noncommon regressors Xnc.}

\item{nb_pts}{the constant C in DGM for the epsilon_0, the lower bound on the grid for epsilon, taken equal to nb_pts*ln(n)/n. Default is 1 without regressors Xc, 3 with Xc.}

\item{Opt}{the option of the method to compute the critical values, either "boot" (numerical bootstrap) or "subsampling". Default is numerical bootstrap.}

\item{lim}{the limit number of observations under which we do no compute the conditional variance.}

\item{weights_x}{the sampling weights for the dataset (Xnc,Xc).}

\item{weights_y}{the sampling weights for the dataset (Y,Xc).}

\item{refs0}{indicating the positions in the vector values corresponding to the components of betac.}

\item{trunc}{equal to 2, for the definition of epsilon.}

\item{boot_par}{the numerical bootstrap parameter. Default is 0.3 without Xc, 0.35 with Xc.}

\item{data_k}{the number of points for the grid search on epsilon. Default is 30. If NULL, then epsilon is taken fixed equal to kp.}

\item{C0}{the upper bound on the grid for epsilon. Default is 0.5.}

\item{c_sign}{sign restrictions on the commonly observed regressors: -1 for a minus sign, 1 for a plus sign, 0 otherwise. Default is NULL, i.e. no constraints.}

\item{nc_sign}{sign restrictions on the non-commonly observed regressors Xnc: -1 for a minus sign, 1 for a plus sign, 0 otherwise. Default is NULL, i.e. no constraints.}

\item{meth}{the method for the choice of epsilon, either "adapt", i.e. adapted to the direction or "min" the minimum over the directions. Default is "adapt".}

\item{nbCores}{number of cores for the parallel computation. Default is 1.}

\item{winsor}{indicates if winsorisation. Default is FALSE.}

\item{version}{version of the computation of the ratio, "first" is a degraded version but fast; "second" is a correct version but slower. Default is "second".}

\item{alpha}{the level for the confidence regions. Default is 0.05.}
}
\value{
a matrix containing the values of the selected epsilon(q) for q directions in sam1.
}
\description{
Function for the data-driven selection of the epsilon tuning parameter
}
