% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabulate_values.R
\name{tabulate_values}
\alias{tabulate_values}
\title{Function to tabulate the values the common regressors Xc whatever the dimension.}
\usage{
tabulate_values(k, values, Xc0, dimXc)
}
\arguments{
\item{k}{the considered value in "values"}

\item{values}{the different unique points of support of the common regressor Xc.}

\item{Xc0}{dataset containing Xc, the common regressors.}

\item{dimXc}{the dimension of Xc}
}
\value{
a matrix of the number of times the kth value in the vector values appears.
}
\description{
Function to tabulate the values the common regressors Xc whatever the dimension.
}
