% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_ratio_variance.R
\name{compute_ratio_variance}
\alias{compute_ratio_variance}
\title{Function to compute the variance bounds}
\usage{
compute_ratio_variance(x, Xp, Yp, dimX2, weights_xp, weights_yp)
}
\arguments{
\item{x}{a matrix containing the directions to compute the variance bounds on the radial function.}

\item{Xp}{the observations of the noncommon regressor (possibly conditional on Xc).}

\item{Yp}{the observations of the outcome variable.}

\item{dimX2}{the dimension of the noncommon regressors Xnc.}

\item{weights_xp}{the sampling or bootstrap weights for the dataset (Xnc,Xc).}

\item{weights_yp}{the sampling or bootstrap weights for the dataset (Y,Xc).}
}
\value{
the value of the ratio of the variance entering the variance bounds.
}
\description{
Function to compute the variance bounds
}
